WIP: modularize graph
